requirements.txt
groq==0.13.0
pandas==2.2.3
python-dotenv==1.0.1
torch==2.5.0
bandit
mypy
ruff